17 Adding stat methods and tests

+++ src/main/java/ru/restaurant/voting/service/restaurant/RestaurantServiceImpl.java
@@ -7,15 +7,19 @@ import org.springframework.util.Assert;
 import ru.restaurant.voting.model.Dish;
 import ru.restaurant.voting.model.Restaurant;
 import ru.restaurant.voting.model.Vote;
+import ru.restaurant.voting.repository.DayMenuRepository;
 import ru.restaurant.voting.repository.DishRepository;
 import ru.restaurant.voting.repository.RestaurantRepository;
 import ru.restaurant.voting.repository.VoteRepository;
-import ru.restaurant.voting.to.RestaurantTo;
+import ru.restaurant.voting.to.RestaurantNamesTo;
+import ru.restaurant.voting.to.RestaurantToWithStats;
 import ru.restaurant.voting.util.exception.NotFoundException;
+import ru.restaurant.voting.util.exception.RestaurantHasNotMenuForThisDay;
 import ru.restaurant.voting.util.exception.UserAlreadyHasVotedException;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -29,12 +33,17 @@ public class RestaurantServiceImpl implements RestaurantService {
     private final RestaurantRepository restaurantRepository;
     private final VoteRepository voteRepository;
     private final DishRepository dishRepository;
+    private final DayMenuRepository dayMenuRepository;
 
     @Autowired
-    public RestaurantServiceImpl(RestaurantRepository restaurantRepository, VoteRepository voteRepository, DishRepository dishRepository) {
+    public RestaurantServiceImpl(RestaurantRepository restaurantRepository,
+                                 VoteRepository voteRepository,
+                                 DishRepository dishRepository,
+                                 DayMenuRepository dayMenuRepository) {
         this.restaurantRepository = restaurantRepository;
         this.voteRepository = voteRepository;
         this.dishRepository = dishRepository;
+        this.dayMenuRepository = dayMenuRepository;
     }
 
     @Override
@@ -66,10 +75,21 @@ public class RestaurantServiceImpl implements RestaurantService {
     }
 
     @Override
-    public List<RestaurantTo> getAll() {
+    public List<RestaurantNamesTo> getAll() {
         return restaurantRepository.getAll().stream()
                 .map(this::asTo)
                 .collect(Collectors.toList());
+    }
+
+    @Transactional
+    @Override
+    public List<RestaurantToWithStats> getAllWithStats() {
+        List<RestaurantToWithStats> all = getAll().stream()
+                .map(RestaurantToWithStats::new)
+                .collect(Collectors.toList());
+        all.forEach(r -> r.setStat(getStat(r.getId())));
+        sortByStatAndName(all);
+        return all;
     }
 
     @Override
@@ -82,6 +102,7 @@ public class RestaurantServiceImpl implements RestaurantService {
         return restaurantRepository.getAllForDay(Objects.requireNonNullElseGet(localDate, LocalDate::now));
     }
 
+    @Transactional
     @Override
     public Vote vote(LocalDate date, int userId, int restaurantId, LocalTime time) {
         if (time == null) {
@@ -90,12 +111,15 @@ public class RestaurantServiceImpl implements RestaurantService {
         if (date == null) {
             date = LocalDate.now();
         }
+        if (dayMenuRepository.getAllForDateAndRestaurantId(restaurantId, date).isEmpty()) {
+            throw new RestaurantHasNotMenuForThisDay(String.format("Restaurant with restaurantId = %d has not DayMenu for this day", restaurantId));
+        }
         Optional<Vote> optionalVote = voteRepository.findByUserIdAndVotingDate(userId, date);
         if (optionalVote.isEmpty()) {
             return voteRepository.save(new Vote(date, userId, restaurantId));
         } else {
             Vote vote = null;
-            if (time.isBefore(LocalTime.of(11, 0,0 ))) {
+            if (time.isBefore(LocalTime.of(11, 0, 0))) {
                 voteRepository.deleteByUserIdAndVotingDate(userId, date);
                 vote = voteRepository.save(new Vote(date, userId, restaurantId));
             } else {
@@ -110,7 +134,41 @@ public class RestaurantServiceImpl implements RestaurantService {
         return dishRepository.getAllDishesByRestaurantId(restaurantId);
     }
 
-    private RestaurantTo asTo(Restaurant restaurant) {
-        return new RestaurantTo(restaurant.getId(), restaurant.getName());
+    @Override
+    public int getStatForDay(LocalDate date, int restaurantId) {
+        if (date == null) {
+            date = LocalDate.now();
+        }
+        return voteRepository.getAllForDate(date, restaurantId).size();
+    }
+
+    @Override
+    public int getStat(int restaurantId) {
+        return voteRepository.getAll(restaurantId).size();
+    }
+
+    @Override
+    public List<RestaurantToWithStats> getAllWithStatsForDay(LocalDate date) {
+        if (date == null) {
+            date = LocalDate.now();
+        }
+        List<RestaurantToWithStats> all = getAllForDay(date).stream()
+                .map(RestaurantToWithStats::new)
+                .collect(Collectors.toList());
+        LocalDate finalDate = date;
+        all.forEach(r -> r.setStat(getStatForDay(finalDate, r.getId())));
+        sortByStatAndName(all);
+        return all;
+    }
+
+    private void sortByStatAndName(List<RestaurantToWithStats> all) {
+        all.sort(Comparator
+                .comparing(RestaurantToWithStats::getStat)
+                .reversed()
+                .thenComparing(RestaurantToWithStats::getName));
+    }
+
+    private RestaurantNamesTo asTo(Restaurant restaurant) {
+        return new RestaurantNamesTo(restaurant.getId(), restaurant.getName());
     }
 }
